requirements.txt
Django
Markdown
django-filter
djangorestframework
sqlparse
validate-docbr